Add AxisSplitter with Enable and Buttons to Rotary selection
Adds modified Axis Splitter that sets output to default values when enable button is not pressed.

Adds Buttons to Rotary selection. Turns two buttons into rotary selection with up to 6 positions with or without rollover.
Would it be possible to separate input and output so user can easily add more buttons to the setup?